W5: Wire FTDIDevice::open() to TACDriveThread; remove IFTDIDriveThread shim
- Replace the `IFTDIDriveThread` virtual interface in AlpacaDevice.h with a
  direct #include of TACDriveThread.h — FTDIDevice now holds a
  `qtac::TACDriveThread*` directly, eliminating the intermediate shim.

- FTDIDevice::open() now lazily loads the .tcnf platform configuration from
  the PlatformContainer entry, then calls setPinSets() on the drive thread
  before start() so the FTDI chipset is opened with the correct bus mask.

- buildMapping() changed from assert to early return when
  _ftdiPlatformConfiguration is null (graceful no-op before config is loaded).

- TACDriveThread base adds setPinSets() no-op default and fixes setPinState
  signature (uint16_t → uint64_t) to match all concrete overrides.

- Build verified: all 24 qtac-core + qt-adapter + qtac-app targets link clean.

+++ source/library/include/qtac/TACPIC32CXDriveThread.h
@@ -66,7 +66,7 @@ public:
 
     void endTransaction(ReceiveInterface* receiveInterface = nullptr);
 
-    void setPinState(uint16_t pin, bool state) override;
+    void setPinState(uint64_t pin, bool state) override;
     void sendCommandSequence(CommandEntries& commandEntries) override;
 
     // PIC32CX does not use reset count.
